refactor: moved contract data tools to tools/contracts
