+++ src/tools/contracts/updateContractData.js
@@ -1,17 +1,17 @@
 import fs from 'fs'
 import path from 'path'
-import { getLatestBlockNumber, getContractData, parseArguments } from './utils.js'
+import { getLatestBlockNumber, getContractData, updateContractData, parseArguments } from '../utils.js'
 
 const toolName = process.argv[1].split('/').pop()
 
 function printUsageAndExit () {
-  console.log(`Usage: npx babel-node src/tools/${toolName} [options]`)
+  console.log(`Usage: npx babel-node src/tools/contracts/${toolName} [options]`)
   console.log(`Options:`)
-  console.log(`  --address <string>   The contract address to fetch data for (required)`)
-  console.log(`  --block <number>     Block number to fetch data at (default: latest block)`)
+  console.log(`  --address <string>   The contract address to update data for (required)`)
+  console.log(`  --block <number>     Block number to update data at (default: latest block)`)
   console.log(`Examples:`)
-  console.log(`  npx babel-node src/tools/${toolName} --address 0x123...`)
-  console.log(`  npx babel-node src/tools/${toolName} --address 0x123... --block 5000000`)
+  console.log(`  npx babel-node src/tools/contracts/${toolName} --address 0x123...`)
+  console.log(`  npx babel-node src/tools/contracts/${toolName} --address 0x123... --block 5000000`)
   process.exit(1)
 }
 
@@ -34,7 +34,7 @@ async function main () {
     const normalizedAddress = options.address.toLowerCase()
 
     console.log(`${toolName}`)
-    console.log(`Fetching data for contract: ${normalizedAddress}`)
+    console.log(`Updating data for contract: ${normalizedAddress}`)
 
     // Get block number if not provided
     let targetBlockNumber = options.blockNumber
@@ -47,6 +47,12 @@ async function main () {
 
     const contractData = await getContractData(normalizedAddress, targetBlockNumber)
 
+    if (!contractData) {
+      throw new Error('Failed to fetch contract data')
+    }
+
+    await updateContractData(contractData)
+
     const result = {
       blockNumber: targetBlockNumber,
       contractData
@@ -54,17 +60,17 @@ async function main () {
 
     console.log('\n=== RESULT ===')
     console.dir(result, { depth: null })
-    console.log(`\nFetch time: ${contractData.fetchTimeMs}ms`)
+    console.log(`\nFetch time: ${contractData.fetchTimeMs}ms, Update time: ${contractData.updateTimeMs}ms`)
 
     // Save result to file
-    const fileName = `contract-data-${normalizedAddress}-${Date.now()}.json`
+    const fileName = `contract-update-${normalizedAddress}-${Date.now()}.json`
     const resultFilePath = path.join(__dirname, fileName)
     fs.writeFileSync(resultFilePath, JSON.stringify(result, null, 2))
     console.log(`\nResults saved to: ${fileName}`)
 
     process.exit(0)
   } catch (error) {
-    console.log(`[${toolName}]: Error fetching contract data`)
+    console.log(`[${toolName}]: Error updating contract data`)
     console.error(error)
     process.exit(1)
   }
